// backend/firebase/functions/src/generated/proto/thing/propeller.ts
export interface Propeller {
  hub: PropellerHub | undefined;
  blades: PropellerBlade[];
}

export interface PropellerHub {
  make: string;
  model: string;
  serial: string;
}

export interface PropellerBlade {
  make: string;
  model: string;
  serial: string;
}

function createBasePropeller(): Propeller {
  return { hub: undefined, blades: [] };
}

export const Propeller: MessageFns<Propeller> = {
  encode(message: Propeller, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.hub !== undefined) {
      PropellerHub.encode(message.hub, writer.uint32(10).fork()).join();
    }
    for (const v of message.blades) {
      PropellerBlade.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Propeller {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePropeller();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.hub = PropellerHub.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.blades.push(PropellerBlade.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Propeller {
    return {
      hub: isSet(object.hub) ? PropellerHub.fromJSON(object.hub) : undefined,
      blades: globalThis.Array.isArray(object?.blades) ? object.blades.map((e: any) => PropellerBlade.fromJSON(e)) : [],
    };
  },

  toJSON(message: Propeller): unknown {
    const obj: any = {};
    if (message.hub !== undefined) {
      obj.hub = PropellerHub.toJSON(message.hub);
    }
    if (message.blades?.length) {
      obj.blades = message.blades.map((e) => PropellerBlade.toJSON(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Propeller>, I>>(base?: I): Propeller {
    return Propeller.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Propeller>, I>>(object: I): Propeller {
    const message = createBasePropeller();
    message.hub = (object.hub !== undefined && object.hub !== null) ? PropellerHub.fromPartial(object.hub) : undefined;
    message.blades = object.blades?.map((e) => PropellerBlade.fromPartial(e)) || [];
    return message;
  },
};

function createBasePropellerHub(): PropellerHub {
  return { make: "", model: "", serial: "" };
}

export const PropellerHub: MessageFns<PropellerHub> = {
  encode(message: PropellerHub, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.make !== "") {
      writer.uint32(10).string(message.make);
    }
    if (message.model !== "") {
      writer.uint32(18).string(message.model);
    }
    if (message.serial !== "") {
      writer.uint32(26).string(message.serial);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PropellerHub {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePropellerHub();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.make = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.model = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.serial = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PropellerHub {
    return {
      make: isSet(object.make) ? globalThis.String(object.make) : "",
      model: isSet(object.model) ? globalThis.String(object.model) : "",
      serial: isSet(object.serial) ? globalThis.String(object.serial) : "",
    };
  },

  toJSON(message: PropellerHub): unknown {
    const obj: any = {};
    if (message.make !== "") {
      obj.make = message.make;
    }
    if (message.model !== "") {
      obj.model = message.model;
    }
    if (message.serial !== "") {
      obj.serial = message.serial;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PropellerHub>, I>>(base?: I): PropellerHub {
    return PropellerHub.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PropellerHub>, I>>(object: I): PropellerHub {
    const message = createBasePropellerHub();
    message.make = object.make ?? "";
    message.model = object.model ?? "";
    message.serial = object.serial ?? "";
    return message;
  },
};

function createBasePropellerBlade(): PropellerBlade {
  return { make: "", model: "", serial: "" };
}

export const PropellerBlade: MessageFns<PropellerBlade> = {
  encode(message: PropellerBlade, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.make !== "") {
      writer.uint32(10).string(message.make);
    }
    if (message.model !== "") {
      writer.uint32(18).string(message.model);
    }
    if (message.serial !== "") {
      writer.uint32(26).string(message.serial);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PropellerBlade {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePropellerBlade();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.make = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.model = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.serial = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PropellerBlade {
    return {
      make: isSet(object.make) ? globalThis.String(object.make) : "",
      model: isSet(object.model) ? globalThis.String(object.model) : "",
      serial: isSet(object.serial) ? globalThis.String(object.serial) : "",
    };
  },

  toJSON(message: PropellerBlade): unknown {
    const obj: any = {};
    if (message.make !== "") {
      obj.make = message.make;
    }
    if (message.model !== "") {
      obj.model = message.model;
    }
    if (message.serial !== "") {
      obj.serial = message.serial;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PropellerBlade>, I>>(base?: I): PropellerBlade {
    return PropellerBlade.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PropellerBlade>, I>>(object: I): PropellerBlade {
    const message = createBasePropellerBlade();
    message.make = object.make ?? "";
    message.model = object.model ?? "";
    message.serial = object.serial ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}